feat(player): unify completion screen
