+++ packages/player/messages/de.po
@@ -57,27 +57,22 @@ msgctxt "usfevO"
 msgid "Keyboard shortcut: press <kbd>{shortcut}</kbd> to choose this answer."
 msgstr "Tastenkürzel: Drücke <kbd>{shortcut}</kbd>, um diese Antwort auszuwählen."
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 msgctxt "E2sGaF"
 msgid "Exit"
 msgstr "Beenden"
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 msgctxt "FazwRl"
 msgid "Try again"
 msgstr "Erneut versuchen"
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 #: src/components/completion-milestone-actions.tsx
 #: src/components/step-navigation-button-group.tsx
 msgctxt "9-Ddtu"
 msgid "Next"
 msgstr "Weiter"
-
-#: src/components/completion-auth-branch.tsx
-msgctxt "odXlk8"
-msgid "Log in"
-msgstr "Anmelden"
 
 #: src/components/completion-brain-power-milestone.tsx
 msgctxt "fS3hB9"
@@ -455,16 +450,6 @@ msgstr "Fortschritt wird nicht gespeichert"
 msgctxt "JDumQX"
 msgid "You can try this lesson without an account, but your score, course progress, levels, and Energy won't be saved."
 msgstr "Du kannst diese Lektion ohne Konto ausprobieren, aber deine Punktzahl, Kursfortschritt, Levels und Energie werden nicht gespeichert."
-
-#: src/components/unauthenticated-progress-prompt.tsx
-msgctxt "rxdOwg"
-msgid "This lesson wasn't saved"
-msgstr "Diese Lektion wurde nicht gespeichert"
-
-#: src/components/unauthenticated-progress-prompt.tsx
-msgctxt "Wamgyz"
-msgid "Log in before your next lesson to keep future progress, scores, levels, and Energy."
-msgstr "Melde dich vor deiner nächsten Lektion an, damit künftige Fortschritte, Punktzahlen, Levels und Energie gespeichert werden."
 
 #: src/components/verdict-label.tsx
 msgctxt "xmF49T"
