+++ packages/player/messages/en.po
@@ -57,27 +57,22 @@ msgctxt "usfevO"
 msgid "Keyboard shortcut: press <kbd>{shortcut}</kbd> to choose this answer."
 msgstr "Keyboard shortcut: press <kbd>{shortcut}</kbd> to choose this answer."
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 msgctxt "E2sGaF"
 msgid "Exit"
 msgstr "Exit"
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 msgctxt "FazwRl"
 msgid "Try again"
 msgstr "Try again"
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 #: src/components/completion-milestone-actions.tsx
 #: src/components/step-navigation-button-group.tsx
 msgctxt "9-Ddtu"
 msgid "Next"
 msgstr "Next"
-
-#: src/components/completion-auth-branch.tsx
-msgctxt "odXlk8"
-msgid "Log in"
-msgstr "Log in"
 
 #: src/components/completion-brain-power-milestone.tsx
 msgctxt "fS3hB9"
@@ -455,16 +450,6 @@ msgstr "Progress won't be saved"
 msgctxt "JDumQX"
 msgid "You can try this lesson without an account, but your score, course progress, levels, and Energy won't be saved."
 msgstr "You can try this lesson without an account, but your score, course progress, levels, and Energy won't be saved."
-
-#: src/components/unauthenticated-progress-prompt.tsx
-msgctxt "rxdOwg"
-msgid "This lesson wasn't saved"
-msgstr "This lesson wasn't saved"
-
-#: src/components/unauthenticated-progress-prompt.tsx
-msgctxt "Wamgyz"
-msgid "Log in before your next lesson to keep future progress, scores, levels, and Energy."
-msgstr "Log in before your next lesson to keep future progress, scores, levels, and Energy."
 
 #: src/components/verdict-label.tsx
 msgctxt "xmF49T"
