+++ packages/player/messages/es.po
@@ -57,27 +57,22 @@ msgctxt "usfevO"
 msgid "Keyboard shortcut: press <kbd>{shortcut}</kbd> to choose this answer."
 msgstr "Atajo de teclado: presiona <kbd>{shortcut}</kbd> para elegir esta respuesta."
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 msgctxt "E2sGaF"
 msgid "Exit"
 msgstr "Salir"
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 msgctxt "FazwRl"
 msgid "Try again"
 msgstr "Inténtalo de nuevo"
 
-#: src/components/completion-auth-branch.tsx
+#: src/components/completion-actions.tsx
 #: src/components/completion-milestone-actions.tsx
 #: src/components/step-navigation-button-group.tsx
 msgctxt "9-Ddtu"
 msgid "Next"
 msgstr "Siguiente"
-
-#: src/components/completion-auth-branch.tsx
-msgctxt "odXlk8"
-msgid "Log in"
-msgstr "Iniciar sesión"
 
 #: src/components/completion-brain-power-milestone.tsx
 msgctxt "fS3hB9"
@@ -455,16 +450,6 @@ msgstr "No se guardará el progreso"
 msgctxt "JDumQX"
 msgid "You can try this lesson without an account, but your score, course progress, levels, and Energy won't be saved."
 msgstr "Puedes probar esta lección sin una cuenta, pero no se guardarán tu puntuación, el progreso del curso, los niveles ni la Energía."
-
-#: src/components/unauthenticated-progress-prompt.tsx
-msgctxt "rxdOwg"
-msgid "This lesson wasn't saved"
-msgstr "Esta lección no se guardó"
-
-#: src/components/unauthenticated-progress-prompt.tsx
-msgctxt "Wamgyz"
-msgid "Log in before your next lesson to keep future progress, scores, levels, and Energy."
-msgstr "Inicia sesión antes de tu próxima lección para conservar tu progreso, puntuaciones, niveles y Energía en el futuro."
 
 #: src/components/verdict-label.tsx
 msgctxt "xmF49T"
